+++ src/fidesops/api/v1/endpoints/oauth_endpoints.py
@@ -76,7 +76,10 @@ async def acquire_access_token(
     else:
         raise AuthenticationFailure(detail="Authentication Failure")
 
-    client_detail = ClientDetail.get(db, object_id=client_id, config=config)
+    # scopes param is only used if client is root client, otherwise we use the client's associated scopes
+    client_detail = ClientDetail.get(
+        db, object_id=client_id, config=config, scopes=SCOPE_REGISTRY
+    )
 
     if client_detail is None:
         raise AuthenticationFailure(detail="Authentication Failure")