+++ src/fidesops/util/oauth_util.py
@@ -21,6 +21,7 @@ from sqlalchemy.orm import Session
 from starlette.status import HTTP_404_NOT_FOUND
 
 from fidesops.api.deps import get_db
+from fidesops.api.v1.scope_registry import SCOPE_REGISTRY
 from fidesops.api.v1.urn_registry import TOKEN, V1_URL_PREFIX
 from fidesops.core.config import config
 from fidesops.models.policy import PolicyPreWebhook
@@ -138,8 +139,9 @@ async def verify_oauth_client(
     if not client_id:
         raise AuthorizationError(detail="Not Authorized for this action")
 
+    # scopes param is only used if client is root client, otherwise we use the client's associated scopes
     client = ClientDetail.get(
-        db, object_id=client_id, config=config, scopes=security_scopes.scopes
+        db, object_id=client_id, config=config, scopes=SCOPE_REGISTRY
     )
 
     if not client: